+++ src/module/config.mjs
@@ -1819,3 +1819,47 @@ export const projects = {
   ],
 };
 preLocalize("projects.types", { key: "label" });
+
+/**
+ * @typedef SourceBook
+ * @property {string} label   An i18n key for the label that will show in sheet headers.
+ * @property {string} name    An i18n key for the longer title that will display in the Compendium Browser.
+ */
+
+/**
+ * Source books provided in the data list for the Document Source Input.
+ * @type {Record<string, SourceBook>}
+ */
+const sourceBooks = {
+  Heroes: {
+    label: "DRAW_STEEL.SOURCE.Books.Heroes.Label",
+    title: "DRAW_STEEL.SOURCE.Books.Heroes.Title",
+  },
+  Monsters: {
+    label: "DRAW_STEEL.SOURCE.Books.Monsters.Label",
+    title: "DRAW_STEEL.SOURCE.Books.Monsters.Title",
+  },
+};
+
+/**
+ * @typedef ContentLicense
+ * @property {string} [label] An i18n key pointing to longer version of the name of the license.
+ */
+
+/**
+ * Licenses for Draw Steel content.
+ * The keys are human readable in the sources panel.
+ * @type {Record<string, ContentLicense>}
+ */
+const sourceLicenses = {
+  "Draw Steel Creator License": {
+    label: "DRAW_STEEL.SOURCE.Licenses.DSCL",
+  },
+};
+
+export const sourceInfo = {
+  books: sourceBooks,
+  licenses: sourceLicenses,
+};
+preLocalize("sourceInfo.books", { keys: ["label", "title"] });
+preLocalize("sourceInfo.licenses", { keys: ["label"] });